fix 별점 ui 별로 하는 기능 수정 중(리뷰 작성, 작성자 일치 아닐 시 표시 완성)

+++ src/main/java/com/multi/culture_link/festival/controller/FestivalController.java
@@ -519,7 +519,7 @@ public class FestivalController {
 	 * @return
 	 */
 	@PostMapping("/insertFestivalReview")
-	public String insertFestivalReview(@RequestParam("reviewTextArea") String reviewTextArea, @RequestParam("reviewStar") double reviewStar, @RequestParam("uploadFile") MultipartFile uploadFile, @RequestParam("festivalId") int festivalId, @AuthenticationPrincipal VWUserRoleDTO user) {
+	public String insertFestivalReview(@RequestParam("reviewTextArea") String reviewTextArea, @RequestParam("reviewStar") int reviewStar, @RequestParam("uploadFile") MultipartFile uploadFile, @RequestParam("festivalId") int festivalId, @AuthenticationPrincipal VWUserRoleDTO user) {
 		
 		try {
 			
@@ -605,7 +605,7 @@ public class FestivalController {
 	 */
 	@PostMapping("/updateFestivalReview")
 	@ResponseBody
-	public String updateFestivalReview(@RequestParam("festivalReviewId") int festivalReviewId, @RequestParam("reviewText") String reviewText, @RequestParam("reviewStar") double reviewStar) {
+	public String updateFestivalReview(@RequestParam("festivalReviewId") int festivalReviewId, @RequestParam("reviewText") String reviewText, @RequestParam("reviewStar") int reviewStar) {
 		
 		try {
